Enhance QEMU boot process and GUI application
- Updated `run-qemu.bat` to use `-cpu max` and log output to `qemu-debug.log`.
- Added detailed CPU state logging to `qemu-debug.log` for better debugging.
- Modified GUI application to increase window height and reduce tab count, focusing on home and changelog tabs.
- Updated changelog content to reflect recent development milestones and features.
- Adjusted boot configuration in `grub.cfg` for direct booting without menu display.

+++ src/intf/idt.h
@@ -14,10 +14,10 @@ typedef struct {
     uint32_t reserved2;
 } __attribute__((packed)) idt_entry_t;
 
-// Structure for the IDT register
+// Structure for the IDT register (64-bit)
 typedef struct {
     uint16_t limit;
-    uint32_t base;
+    uint64_t base;
 } __attribute__((packed)) idt_ptr_t;
 
 // Function to set an IDT entry
